Whoops.  Changed behavior so size argument is bytes and not bits.
git-svn-id: https://fluks.googlecode.com/svn@71 6b01645e-0d02-11de-a42d-873b86d837ff

+++ whirlpool.h
@@ -47,7 +47,7 @@ __BEGIN_DECLS
 
 void	whirlpool_init(struct whirlpool_ctx *const ctx);
 void	whirlpool_update(struct whirlpool_ctx *const ctx,
-	    const uint8_t *const buf, unsigned long sz);
+	    const uint8_t *const buf, size_t sz);
 void	whirlpool_end(struct whirlpool_ctx *const ctx, uint8_t *const buf);
 
 __END_DECLS
